+++ changelog.md
@@ -24,6 +24,7 @@
       - When using arrow or numpad to move around map, it would not correctly scroll the newly selected room if you have scrolled off the old room
       - Fixed mouse hover room while scroll, as old x/y they are now invalid
       - When after editing a room property and clicking on a read only field it would allow you to edit the field
+      - Fixed advanced editors for single lines to wrap text
     - Fixed issue when split view and scrolling tabs not correctly setting scroll width
     - String blocks where not correctly terminated and formatted text in side
     - Save file as would not correctly clear the previous file name data